LayoutTests/fast/url/scheme-expected.txt

Canonicalization of URL schemes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS canonicalize('http://example.com/') is 'http://example.com/'
FAIL canonicalize('HTTP://example.com/') should be http://example.com/. Was HTTP://example.com/.
PASS successfullyParsed is true

TEST COMPLETE
